[Update] Completed saving all information except status to order_detail

+++ app/controllers/public/orders_controller.rb
@@ -34,23 +34,23 @@ class Public::OrdersController < ApplicationController
 
   def create
     @order = Order.new(order_params)
-    @order.save
-    # @order.customer_id = current_customer.id
-    @cart_items = current_customer.cart_items.all
-
-    @cart_items.each do |cart_item|
-      @order_detail = OrderDetail.new
-      @order_detail.order_id = @order.id
-      @order_detail.item_id = cart_item.item.id
-      @order_detail.price = cart_item.item.price
-      @order_detail.amount = cart_item.amount
-      @order_detail.save
-    end
-
-    CartItem.destroy_all
-    redirect_to orders_complete_path
-  end
+    @order.customer_id = current_customer.id
+    if @order.save
+      @cart_items = current_customer.cart_items.all
+      @cart_items.each do |cart_item|
+        @order_detail = OrderDetail.new
+        @order_detail.order_id = @order.id
+        @order_detail.item_id = cart_item.item.id
+        @order_detail.price = cart_item.item.price
+        @order_detail.amount = cart_item.amount
+        @order_detail.save
+      end
   
+      CartItem.destroy_all
+      redirect_to orders_complete_path
+    end
+  end
+
   def complete
   end
